feat: create ConfirmAction component which asks for user confirmation before folder deletion

+++ src/components/Folder.jsx
@@ -2,12 +2,16 @@ import React, { useState } from 'react';
 import dots from '../assets/images/dots.svg';
 import { useDeleteFolder } from './contexts/StorageContext';
 import FolderOptions from './FolderOptions';
+import ConfirmAction from './ConfirmAction';
 
 function Folder({ name, active, makeActive, id }) {
+	const [showConfirm, setShowConfirm] = useState(false);
 	const [showOptions, setShowOptions] = useState(false);
 	const displayOptions = () => setShowOptions(true);
 	const hideOptions = () => setShowOptions(false);
 	const handleDelete = useDeleteFolder();
+
+	const toggleConfirm = () => setShowConfirm((prev) => !prev);
 
 	return (
 		<div
@@ -20,7 +24,14 @@ function Folder({ name, active, makeActive, id }) {
 			{showOptions && (
 				<FolderOptions
 					hideOptions={hideOptions}
+					toggleConfirm={toggleConfirm}
+				/>
+			)}
+			{showConfirm && (
+				<ConfirmAction
 					handleDelete={() => handleDelete(id)}
+					name={name}
+					toggleConfirm={toggleConfirm}
 				/>
 			)}
 		</div>
